clk: qcom: clk-cpu-osm: Query current CPU frequency from OSM
Instead of getting the index and consulting the in-kernel frequency
table for the frequency corresponding to the fetched index, read the XO
clock multiplier for the current frequency directly from OSM and
multiply it by the XO rate. This ensures that if the kernel frequency
table has been altered, accurate readings can still be obtained.

300 MHz (OSM_INIT_RATE) is handled separately due to the use of a
separate clock source. All other frequencies are derived from the 19.2
MHz XO clock.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: CloudedQuartz <ravenklawasd@gmail.com>