drivers: thermal: step_wise: Handle the temperature stable trend
Step wise algorithm ignores the temperature stable trend, which can
cause a case where step wise will not clear the mitigation, when the
temperature is lower than hysteresis and stays stable throughout.

In order to avoid this, consider temperature stable trend similar to
temperature dropping trend. This will lower the mitigation when the
temperature is less than hysteresis.

Change-Id: Ic7d1d35c74d732562193195e46af945ef41912c2
Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: DennySPB <dennyspb@gmail.com>
Signed-off-by: PainKiller3 <ninadpatil100@gmail.com>
Signed-off-by: CloudedQuartz <ravenklawasd@gmail.com>